+++ src/one_char.rs
@@ -23,17 +23,25 @@ pub struct OneChar(pub char);
 
 impl OneChar {
     pub fn description(self) -> String {
+        let scalar_value = u32::from(self.0);
         let mut utf8 = vec![0; self.0.len_utf8()];
         self.0.encode_utf8(utf8.as_mut_slice());
-        let utf8_hex = utf8.iter().map(|byte| format!("{:x}", byte)).collect::<String>();
-        format!("{} [U+{:04X}, {}, 0x{}]", self.0, u32::from(self.0), u32::from(self.0), utf8_hex)
+        let bytes = utf8.iter().map(|byte| format!("0x{:X}", byte)).collect::<Vec<_>>();
+        format!(
+            "\"{}\"(U+{:0scalar_value_length$X}, {}, {})",
+            self.0,
+            scalar_value,
+            scalar_value,
+            bytes.join(" "),
+            scalar_value_length = if scalar_value > 0xFFFF { 6 } else { 4 },
+        )
     }
 }
 
 #[derive(Debug, Copy, Clone, Eq, PartialEq, Hash)]
 pub enum ParseError {
     EmptyInput,
-    InvalidCodePoint(u32),
+    InvalidUnicodeScalarValue(u32),
     InvalidDigitInRadix(u32),
     UTF8BytesEmpty,
     UTF8BytesStrCantAlignToBytes,
@@ -45,7 +53,9 @@ impl ToString for ParseError {
     fn to_string(&self) -> String {
         match self {
             Self::EmptyInput => "you don't input anything".to_string(),
-            Self::InvalidCodePoint(codepoint) => format!("invalid unicode codepoint {}", codepoint),
+            Self::InvalidUnicodeScalarValue(scalar_value) => {
+                format!("invalid unicode scalar value {}", scalar_value)
+            }
             Self::InvalidDigitInRadix(radix) => {
                 format!("has invalid char when parse as a base {} number", radix)
             }
@@ -60,30 +70,32 @@ impl ToString for ParseError {
 }
 
 impl OneChar {
-    pub fn from_codepoint(codepoint: u32) -> Result<Self, ParseError> {
-        char::try_from(codepoint).map(Self).map_err(|_| ParseError::InvalidCodePoint(codepoint))
+    pub fn from_scalar_value(scalar_value: u32) -> Result<Self, ParseError> {
+        char::try_from(scalar_value)
+            .map(Self)
+            .map_err(|_| ParseError::InvalidUnicodeScalarValue(scalar_value))
     }
 
-    pub fn from_codepoint_str_radix(s: &str, radix: u32) -> Result<Self, ParseError> {
-        Self::from_codepoint(
+    pub fn from_scalar_value_str_radix(s: &str, radix: u32) -> Result<Self, ParseError> {
+        Self::from_scalar_value(
             u32::from_str_radix(s, radix).map_err(|_| ParseError::InvalidDigitInRadix(radix))?,
         )
     }
 
-    pub fn from_codepoint_str_bin(s: &str) -> Result<Self, ParseError> {
-        Self::from_codepoint_str_radix(s, 2)
+    pub fn from_scalar_value_str_bin(s: &str) -> Result<Self, ParseError> {
+        Self::from_scalar_value_str_radix(s, 2)
     }
 
-    pub fn from_codepoint_str_oct(s: &str) -> Result<Self, ParseError> {
-        Self::from_codepoint_str_radix(s, 8)
+    pub fn from_scalar_value_str_oct(s: &str) -> Result<Self, ParseError> {
+        Self::from_scalar_value_str_radix(s, 8)
     }
 
-    pub fn from_codepoint_str_dec(s: &str) -> Result<Self, ParseError> {
-        Self::from_codepoint_str_radix(s, 10)
+    pub fn from_scalar_value_str_dec(s: &str) -> Result<Self, ParseError> {
+        Self::from_scalar_value_str_radix(s, 10)
     }
 
-    pub fn from_codepoint_str_hex(s: &str) -> Result<Self, ParseError> {
-        Self::from_codepoint_str_radix(s, 16)
+    pub fn from_scalar_value_str_hex(s: &str) -> Result<Self, ParseError> {
+        Self::from_scalar_value_str_radix(s, 16)
     }
 
     pub fn from_utf8_bytes(s: &str) -> Result<Self, ParseError> {
@@ -99,8 +111,8 @@ impl OneChar {
                 let c1 = c1.unwrap(); // at least one char because of the `take_while`
                 let c2 = c2.ok_or(ParseError::UTF8BytesStrCantAlignToBytes)?;
                 if c1.is_ascii_hexdigit() && c2.is_ascii_hexdigit() {
-                    #[allow(clippy::cast_possible_truncation)]
                     // because two digit hex meets u8 type
+                    #[allow(clippy::cast_possible_truncation)]
                     Ok((c1.to_digit(16).unwrap() << 4 | c2.to_digit(16).unwrap()) as u8)
                 } else {
                     Err(ParseError::InvalidDigitInRadix(16))
@@ -134,11 +146,11 @@ impl FromStr for OneChar {
                 let c1 = c1.to_ascii_lowercase();
                 let c2 = c2.to_ascii_lowercase();
                 match (c1, c2) {
-                    ('0', 'b') => Self::from_codepoint_str_bin(&s[2..]),
-                    ('0', 'o') => Self::from_codepoint_str_oct(&s[2..]),
-                    ('u', '+') => Self::from_codepoint_str_hex(&s[2..]),
+                    ('0', 'b') => Self::from_scalar_value_str_bin(&s[2..]),
+                    ('0', 'o') => Self::from_scalar_value_str_oct(&s[2..]),
+                    ('u', '+') => Self::from_scalar_value_str_hex(&s[2..]),
                     ('0', 'x') => Self::from_utf8_bytes(&s[2..]),
-                    _ => Self::from_codepoint_str_dec(s),
+                    _ => Self::from_scalar_value_str_dec(s),
                 }
             }
             // Because an iterator can not produce value after returning None