+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## WIP
 
+- Remove dependency of `unicode-width` due to only output font family name in english
+- Adjust character output format
+
 ## 0.1.0
 
 - Init release, has same features as [which_fonts_support][which_fonts_support-github] script